vb script to keep the session alive
Download the file to a directory.

Right click + run or double click to execute

Terminate in task manager or just by creating a "New Text Document.txt" in same directory. 

Remember to delete the "New Text Document.txt" is exists before executing the script, if not script will identify the file and terminate after 180secs.